Add: First version of forward button TESTING.
Extended check_button() and run it in a thread.
It waits for differently timed button presses.
Pressing the button once moves to the next song
in the playlist.
Pressing twice inbetween one second advances the
current song 30 s forward.
Holding the button for longer than one second skips
to the next artist.

Added next_song(), previous_song() and seekcur_song()
which are the same methods as in MPD.
Added next_artist() which scans the playlist and plays
the song of the first artist different from the current.