Change: made shutdown() more robust TESTING.
Added thread communication, so shutdown()
does keep the LEDs turned off after it
played its goodbye animation.
This doesn't work though when any LONG_SONG is playing
because of their visualization thread.
Added stop() as a workaround, but that introduces
the inconsistency that playback will not auto resume
after the next startup.