Added code to save and restore the playback state beween restarts TESTING.
I'm still struggling with a clean shutdown(). If the player is to be
turned off, it should always be playing the shutdown animation and keep
the LEDs turned off afterwards.
So i added stop() a few commits ago which stops the playback before
the animation is played. What was missing then was code to restore the
previous state after startup.

Added save_state() and restore_state().
Moved calls to player initialisation to setup().