Changed the way most required threads are started.
They were previously started successively at the beginning of main.
Defined a dict in run["threads"] which holds a token as the
key, the method to be threaded and the thread itself after it's been
started.
Added start_threads() which starts all or specific threads.
Added monitor_threads() which check if the threads are running and
restarts them if necessary.